+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Change github footer link to the tickets repository [#80](https://github.com/etalab/udata-front/pull/80)
 - Remove banner for the new search beta on datasets search page [#83](https://github.com/etalab/udata-front/pull/83)
 - Fix RGAA criterion 7.3 [#82](https://github.com/etalab/udata-front/pull/82)
+- Use avatar_url for owner [#84](https://github.com/etalab/udata-front/pull/84)
 
 ## 1.2.3 (2022-01-27)